Factorize the zx units tests a bit
We want to use peer_unit rather than peer_zx but we still
need the functions for writing strings to the screen and
clearing lines.  These functions have been move into
util_print_zx so that they can be re-used by both the unit
tests and the real zx applications.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>