Fix installation script
The installation script failed if the /BIN/EXTRA directory was not
present on the user's spectrum, which is likely to be the case if
they have never previously installed specasm, meaning that it
was practically impossible for new users to install specasm.  Not
ideal!

This commit fixes the issue by checking for the presence of
/BIN/EXTRA in the install script.  If the directory is not present
it is created by the install script.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>