Remove label subtraction from Specasm
Label subtraction was deprecated in version v5 of Specasm and
is now removed in v7 to free up space needed for the next opcodes.
It's still supported by the linker however, so .x files created
with pre v7 versions of Specasm can still be linked with the v7
linker.  The Specasm editor will rewrite instructions that use
label subtraction to use the new expression syntax when it loads
.x files.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>