Fix specasm .dotn file
The specasm .dotn file now works properly on the next.  This
is achieved by relocating some of the code into 3 separate 8kb
banks, 2 for the parsing code and 1 for the dumping code.  We
have about 1.5kb free in the main section.  We also fix the
keyboard input (for all clockspeeds) and the printing routines
which relied on the ROM being mapped and so needed to be
rewritten.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>