salink: link .x files in ascending order of their names
It is now possible for the user to control the order in which the
code in the various .x files that make up a Specasm project gets
written into the final binary.  The file with the .Main label
always comes first followed by the code in the remainining files
which are written in ascending order of their filenames.  This
feature is intended for programatically generated .x files like
compilers, which might generate more code that can fit into a
single .x file.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>