Allow '-' and '+' to include directories
In addition to be able to add individual .x files to the binary,
'-' and '+' can now be used to include all the .x files in a
given directory.  This makes it much easier to split the program
up into separate directories.  Directory inclusion is not recursive.
If the included directory contains a sub-directory with .x files
these files will be ignored by the linker unless that subdirectory
is also explicitly included.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>