Fix error handling in peer layers
We were not detecting errors correctly in the Spectrum
peer layers.  This resulted in the code that handles
file not found errors in the editor from triggering,
which was a good thing as it was broken.  So we fix the
incorrect error detection in the peer layer and we fix
the editor.  This then fixes a long standing bug where
the contents of the editor are wiped when trying to
open a non-existent file.  Now they are preserved.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>