Finally fix keyboard handling
There were two outstanding issues:

- On my Toastrack, reading the key directly after waitkey() returns
was somethings returning the wrong result for some of the new keys
like Delete or the arrow keys.  Adding a small delay fixes this.
- We had some key specific rules to handle the case where phantom
keys were generated after a shifted key, but these did not cover
all cases.  So now we have a generic solution that converts the
returned key to a scancode and masks off the shift keys, so we can
detect that DELETE followed by '0' are probably the same key.

Hopefully, this is fixed now once and for all.  Seems to work
well now on a KS2 Next, on a Toastrack and on a 48Kb Issue 3.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>