specasm: Add clipboard support on the Next
On the Next the commands 'x', 'cc' and 'v', cut, copy and
paste to and from a clipboard.  This is much less clunky than
the existing move and copy mechanism and allows you to copy
code between different files for the first time.  The 'x'
command can also be used in selecting mode to cut the visibly
selected text to the clipboard.

Also fix a small UI issue when the original copy command fails
due to lack of space in the current file.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>