salink: Increase MAX_FILE_SIZE to 42
This is necessary to provide enough space for including files with
the + directive as the /specasm/ has to be prepended and a .x has
to be potentially appened.  This will reduce the maximum number
of pending files from 35 to 23 when processing includes, but hopefully
that will be enough.  Without this change the + directive is fairly
useless.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>