Add support for unit tests
Add unit test support.  Specasm now supports .t files which can include
unit test code.  This code is included in a separate binary with a .tst
suffix created by the linker.  The .tst file contains a jump table to
all the test functions (functions that begin with a Test prefix) and
the contents of the .t files as well as the contents of all the .x files.

Samake has been modified to generate BASIC test harnesses that invoke
all the Test* functions found in the .tst binary printing their results
to the screen.

.t files are not included when directories are included using - or +
but can be explicitly included by name.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>