specasm: Fix incorrect error code in ld
The ld instruction was reporting the wrong error code when used
with labels or expressions and the number of strings in the .x
were exhausted.  The user should have seen a "Too many strings"
error, but instead they were confronted with a "Bad register"
error which was very confusing.  Should all be fixed now.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>