specasm, tests: port to the 128 spectrum
There's now a 128 version of the Specasm editor.  The 128 version of
Specasm is faster than the 48 version as none of its code or data
(apart from the clipboard) is resides in contended memory.

The Next version of Specasm has also been updated to use 16kb banks
rather than 8kb banks.  This unifies the paging mechanism used by
the 128 and next versions of Specasm, and so simplifies the code,
but also increases the size of the clipboard on the Next version from
8 to 16kb, which should be large enough to store the entire contents
of any file.

Finally, we add a unitest tap file for the 128kb Spectrum and modify
the existing Next unittest program to use 16kb banks, to be
consistent with the editor.

salink still uses 8kb banks on the Next and there's no 128 kb version
yet.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>