specasm: redo the 128 bootstrap
We now have a custom 128 bootstrap helper that contains the
calibration code and some bankswitching code.  The bank switching
code in the bootstrap helper and in specasm_trampolines_128.c
should be Amstrad friendly, using non-contended memory banks
on both Amstrad and Sinclair machines (although this has only
been tested on Sinclair machines and Harlequin based clones).
The location of the keyboard calibration parameter has been
changed to be just above the top of the stack.  The old
address seems to hang Amstrad machines.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>