specasm,saimport: fix nasty formatting issues
As the editor re-formats entered lines, indenting instructions and
adding spaces after commas, it's possible for the user to enter
an instruction that when formatted is > 32 characters, and so
cannot be displayed by the editor.  This was overflowing memory.
We fix the bug by checking that the length of formatted lines
are <= 32 characters, preventing the user for entering the line
if its formatted length is too long.  We also fix another issue
with the formatting of short comments used in long lines.

saimport is also modified on non ZX platforms to check that the
formatted length of a parsed line is <= 32 characters, returning
an error if it is not.  Unfortunately, the ZX versions of
saimport cannot perform this check as there's not enough
memory free, so for now a warning has been added to the
documentation.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>