specasm: allow single letter .t files
Previously, it was impossible to save a .t file whose filename
contained a single letter, excluding the extension, e.g.,
w.t.  It was also not possible to save a .x file with a single
letter when explicitly specifying the extension, e.g,

> s w.x

Both this issues have now been fixed.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>