salink: fix trampolines on the Next
Make sure we always return to the calling bank once the called
function has returned.  Previously, the bank mapped back in
after the function had returned was hard coded which was very
fragile and resulted in a bug in version v9 of specasm which
could cause memory corruption during linking.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>