salink: fix sdcard corruption on linker error
It seems that closing and deleting an empty file causes some
sort of sd card corruption, which prevents the SD card from
subsequently being mounted on macos without being repaired.
The Next doesn't have this issue.  So on ESXDOS we now ensure
that the binary the linker is creating when an error occurs
has at least one byte written to it before we close and
delete it.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>