test: fix the regcheck unit tests
The format of the error message from TstCheckRegs had changed breaking
some of the unit tests.  Also a lot of the tests didn't disable
interrupts and save and restore iy and the prime registers, which
led to crashes and weird behaviour.  I think I must have originally
just tested this on the Next.  The tests all work on a Spectrum
now.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>