specasm: add documentation
A new command 'h' has been added that can be used to bring up
some information about the z80 instructions and assembler
directives.  'h' can be used with or without a parameter.
Without a parameter it displays the help for the first instruction
'adc'.  The parameter when specified is expected to by a z80
instruction or an assembler directive. The command will attempt
to display help on the specified parameter, e.g., 'h in'.  When
in the help screen the arrow keys can be used to move left and
right through the various help screens and any other key returns
to the editor.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>