specasm: add documentation for the Next
The invovles splitting up the documentation strings into separate
files so that they can be stored in separate pages in the Next.  On
the spectrum 128 the documentation code and all the instruction
data is stored in a single 16kb bank.  On the next the code and
all the instruction data apart from the descriptions are stored in
one 8kb bank.  The descriptions are split over 2 8kb banks which
are paged in as necessary.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>