specasm: fix a bug in the help command
The help command with no argument was supposed to take the user
to the first instruction, adc.  This didn't work however as the
string passed to the specasm_help function was stored in the
editor bank and swapped out when the help function was called.
Now we store it on the stack.

A bit of code reformatting is done as well.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>