samac: reduce org address of SAMAC
We could use an ORG address of 32768 for the SAMAC binary, but
this would be cutting it a bit fine, leaving only 500 or so bytes
free.  If the code grew in a subsequent release we'd end up having
to lower the ORG address and this would break users' existing macros.
So let's leave ourselves some room.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>