specasm: fix a paste crash
Pasting a block of code containing long strings into a .x file that
has no long strings available crashes the spectrum on the latest
version of Specasm.  This is because the code mistakenly assumed
that an error could not occur and didn't handle the failure from
specasm_parse_line_e.  We do now trap, handle and report the error.
We stop inserting lines after the first error occurs.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>