release: combine 128 and 48 zip files
We'll now only distribute the 128kb zip file, which contains the
install script and the Specasm binary for the 48kb machine as well
as for the 128kb Sectrum.  The user will need to run the appropriate
install script to install the version that best suits their machine,
INST48.BAS for the 48kb machine and INST128.BAS for the Spectrum
128.

We remove the release creation code for the 48kb Spectrum.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>